Update Filtering for all Users and save to file for both project and Individual Users file
So for filtering everytime u filter whether to view all project or apply for project basically anything to do with viewing all Projects for Offcier for Manager for Applicant, the Filter will remain on once u set the filter, so to unset the filter u have to set everything off. And for the alphabatical order, once u view proj, it will auto sort based on project name.